all effects spread around properly
