+++ src/Cooked/MockChain/Direct.hs
@@ -59,26 +59,6 @@ import PlutusLedgerApi.V3 qualified as Api
 --
 -- - emits entries in a 'MockChainBook'
 
--- | This represents elements that can be emitted throughout a 'MockChain'
--- run. These elements are either log entries corresponding to internal events
--- worth logging, or aliases for hashables corresponding to elements users
--- wishes to be properly displayed when printed with
--- 'Cooked.Pretty.Class.PrettyCooked'
-data MockChainBook where
-  MockChainBook ::
-    { -- | Log entries generated by cooked-validators
-      mcbJournal :: [MockChainLogEntry],
-      -- | Aliases stored by the user
-      mcbAliases :: Map Api.BuiltinByteString String
-    } ->
-    MockChainBook
-
-instance Semigroup MockChainBook where
-  MockChainBook j a <> MockChainBook j' a' = MockChainBook (j <> j') (a <> a')
-
-instance Monoid MockChainBook where
-  mempty = MockChainBook mempty mempty
-
 -- | A 'MockChainT' builds up a stack of monads on top of a given monad @m@ to
 -- reflect the requirements of the simulation. It writes a 'MockChainBook',
 -- updates and reads from a 'MockChainState' and throws possible
